requirements.txt
fastapi
uvicorn
twilio
openai>=1.0
langchain
langchain-openai
langchain-community
python-dotenv
python-multipart
redis            # (opcional, solo si usarás Redis)
chromadb